+++ tests/point_tracker.h
@@ -21,16 +21,16 @@ using namespace cv;
 
 struct	tracking_data
 {
-		float grades_from_center_x;
-		float grades_from_center_y;
+	double grades_from_center_x;
+	double grades_from_center_y;
 };
 
 
-tracking_data get_offset_angles(int webcam_model_number, float image_ratio, int image_width, int image_height, Point position);
+tracking_data get_offset_angles(int webcam_model_number, double image_ratio, int image_width, int image_height, Point position);
 
 tracking_data get_offset_angles(int webcam_model_number, Point position);
 
-float determine_offset_angle(int position, float field_of_view, int number_of_pixels);
+double determine_offset_angle(int position, double field_of_view, int number_of_pixels);
 
 
 #endif //POINT_TRACKER_H
